+++ nat/nat.tf
@@ -38,8 +38,11 @@ resource "aws_route_table" "main" {
   }
 }
 
+# route tables for public subnet, e.g. if we want VPC peering between stacks without NATs
+# running the web process in a public subnet
 resource "aws_route_table_association" "main" {
-  count          = length(var.subnet_private_ids)
-  subnet_id      = var.subnet_private_ids[count.index]
-  route_table_id = aws_route_table.main[count.index].id
+  for_each  = { for idx, subnet in var.subnet_private_ids : subnet => idx }
+  subnet_id = each.key
+
+  route_table_id = aws_route_table.main[each.value].id
 }