+++ rds/variables.tf
@@ -79,6 +79,26 @@ variable "enable_replication" {
   default     = false
 }
 
+variable "subnet_group_name" {
+  type    = string
+  default = null
+}
+
+variable "delete_automated_backups" {
+  type    = bool
+  default = true
+}
+
+variable "skip_final_snapshot" {
+  type    = bool
+  default = false
+}
+
+variable "final_snapshot_identifier" {
+  type    = string
+  default = null
+}
+
 locals {
   name             = var.name != null ? var.name : "${var.project}-${var.environment}${var.regional ? "-${var.region}" : ""}"
   name_underscored = var.name != null ? replace(var.name, "-", "_") : "${var.project}_${var.environment}${var.regional ? "_${var.region}" : ""}"