+++ stack/app/variables.tf
@@ -93,6 +93,11 @@ variable "kms_deletion_window_in_days" {
   type = number
 }
 
+variable "grant_access_to_kms_arns" {
+  type    = list(string)
+  default = []
+}
+
 variable "kms_app_arn" {
   type = string
 }
@@ -132,6 +137,11 @@ variable "remote_cidr_blocks" {
 variable "skip_elasticache" {
   type    = bool
   default = false
+}
+
+variable "elasticache_name" {
+  type    = string
+  default = null
 }
 
 variable "elasticache_node_type" {
@@ -246,6 +256,32 @@ variable "rds_allow_from_cidr_blocks" {
   type    = list(string)
   default = []
 }
+
+variable "rds_subnet_group_name" {
+  type    = string
+  default = null
+}
+
+variable "rds_multi_az" {
+  type        = bool
+  description = "if no value is passed, multi-az will be set to true if the environment is production"
+  default     = null
+}
+
+variable "rds_delete_automated_backups" {
+  type    = bool
+  default = true
+}
+
+variable "rds_skip_final_snapshot" {
+  type    = bool
+  default = false
+}
+
+variable "rds_final_snapshot_identifier" {
+  type    = string
+  default = null
+}
 # =============== RDS ================ #
 
 # =============== ECS ================ #
@@ -355,6 +391,11 @@ variable "cloudwatch_dashboard_view" {
   default = "simple"
 }
 
+variable "enable_cloudwatch_dashboard" {
+  type    = bool
+  default = true
+}
+
 variable "cloudwatch_sns_topic_arns" {
   type    = list(string)
   default = []
@@ -415,5 +456,10 @@ variable "db_instance_class_memory_in_gb" {
 variable "datapoints_to_alarm" {
   type    = number
   default = 1
+}
+
+variable "cloudwatch_logs_retention_in_days" {
+  type    = number
+  default = 90
 }
 # =============== Cloudwatch ================ #