+++ stack/app/vpc-peering.tf
@@ -1,10 +1,6 @@
-locals {
-  enable_vpc_peering = var.vpc_peering || var.rds_is_read_replica
-}
-
 module "vpc-peering" {
   source = "../../vpc-peering"
-  count  = local.enable_vpc_peering ? 1 : 0
+  count  = var.vpc_peering ? 1 : 0
 
   providers = {
     aws      = aws
@@ -14,10 +10,14 @@ module "vpc-peering" {
   project     = var.project
   environment = var.environment
 
-  requester_region              = var.region
-  requester_vpc_id              = module.vpc.id
-  requester_cidr_block          = var.vpc_cidr_block
-  requester_nat_route_table_ids = module.nat.aws_route_table_ids
+  requester_region     = var.region
+  requester_vpc_id     = module.vpc.id
+  requester_cidr_block = var.vpc_cidr_block
+  requester_nat_route_table_ids = flatten(concat(
+    module.nat.aws_route_table_ids,
+    module.vpc.route_table_ids,
+
+  ))
 
   accepter_region              = var.rds_master_db_region
   accepter_vpc_id              = var.rds_master_db_vpc_id